% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Verifica dataframe}
\usage{
check_dataframe(df)
}
\arguments{
\item{df}{Dataframe a ser verificado}
}
\value{
Dataframe vazio.
}
\description{
Verifica se um determinado daframe dado é nulo ou vazio.
}
