% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{.get_tweets_30_days}
\alias{.get_tweets_30_days}
\title{Retorna tweets recentes dos últimos 30 dias}
\usage{
.get_tweets_30_days(q, from_date, to_date, token, env_name)
}
\arguments{
\item{q}{Query da busca}

\item{from_date}{Data mais antiga a ser pesquisada (formato YYYYMMDDHHMM)}

\item{to_date}{Data mais recente a ser pesquisada (formato YYYYMMDDHHMM)}

\item{token}{Token de acesso a requisições PREMIUM}

\item{env_name}{Nome do enviroment (cadastrado no twitter developer)}
}
\value{
Dataframe com tweets
}
\description{
Retorna tweets dos últimos 30 dias que contêm termo passado
}
