% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{generate_token}
\alias{generate_token}
\title{Gera o token de acesso a requisições PREMIUM}
\usage{
generate_token(
  app_name,
  access_token,
  access_token_secret,
  api_key,
  api_secret_key
)
}
\arguments{
\item{app_name}{Nome do aplicativo no Twitter}

\item{access_token}{Access token}

\item{access_token_secret}{Access token secret}

\item{api_key}{API Key}

\item{api_secret_key}{API Secret Key}
}
\value{
Token de acesso a requisições do tipo PREMIUM
}
\description{
A partir das chaves de acesso, gera o token necessário para fazer as requisições do tipo PREMIUM.
}
