% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{generate_twitter_trends}
\alias{generate_twitter_trends}
\title{Gera estatísticas dos termos no twitter}
\usage{
generate_twitter_trends(tweets)
}
\arguments{
\item{tweets}{Dataframe com tweets}
}
\value{
Dataframe com estatísticas dos termos
}
\description{
Cria um dataframe com 'desempenho' dos termos no twitter por semana
}
