% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tweets}
\alias{get_tweets}
\title{Retorna tweets recentes}
\usage{
get_tweets(word)
}
\arguments{
\item{word}{Palavra a ser pesquisada}
}
\value{
Dataframe com tweets
}
\description{
Retorna tweets dos últimos 6-9 dias que contêm termo passado
}
