% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tweets_pls}
\alias{get_tweets_pls}
\title{Retorna tweets recentes}
\usage{
get_tweets_pls(words_df)
}
\arguments{
\item{words_df}{Dataframe com palavras a serem pesquisadas}
}
\value{
Dataframe com tweets
}
\description{
Retorna tweets dos últimos 6-9 dias que contêm termos passados
}
