% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tweets_pls_last_30_days}
\alias{get_tweets_pls_last_30_days}
\title{Retorna tweets recentes filtrados por autores, palavras-chave e intervalo de datas}
\usage{
get_tweets_pls_last_30_days(
  authors_query,
  words_filter,
  from_date,
  to_date,
  token,
  env_name
)
}
\arguments{
\item{authors_query}{Query com os autores a serem pesquisados}

\item{words_filter}{Query com as palavras-chave que filtrarão o conteúdo do tweet}

\item{from_date}{Data mais antiga a ser pesquisada (formato YYYYMMDDHHMM)}

\item{to_date}{Data mais recente a ser pesquisada (formato YYYYMMDDHHMM)}

\item{token}{Token de acesso a requisições PREMIUM}

\item{env_name}{Nome do enviroment (cadastrado no twitter developer)}
}
\value{
Dataframe com tweets filtrados
}
\description{
Retorna tweets dos últimos 30 dias que contêm termos e autores específicos
}
